generator/requirements.txt
azure-eventhub==5.12.0
python-dotenv==1.0.0
